+++ Opus.Models/DbModels/StaffEquipment.cs
@@ -12,7 +12,7 @@ namespace Opus.Models.DbModels
     {
         [Key]
         public int Id { get; set; }
-        public int StaffId { get; set; }
+        public long StaffId { get; set; }
         public long ProductId { get; set; }
         public int Quantity { get; set; }
         public DateTime DeliveryDate {get;set;}
